+++ src/dashboardWebView/DashboardMessage.ts
@@ -17,4 +17,5 @@ export enum DashboardMessage {
   deleteMedia = 'deleteMedia',
   insertPreviewImage = 'insertPreviewImage',
   updateMediaMetadata = 'updateMediaMetadata',
+  createMediaFolder = 'createMediaFolder',
 }