+++ src/dashboardWebView/components/Header/Folders.tsx
@@ -1,4 +1,4 @@
-import { Menu, Transition } from '@headlessui/react';
+import { Menu } from '@headlessui/react';
 import * as React from 'react';
 import { useRecoilState } from 'recoil';
 import { FolderAtom } from '../../state';