+++ src/dashboardWebView/components/Media/Pagination.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { DashboardMessage } from '../../DashboardMessage';
 import { LoadingAtom, MediaTotalSelector, PageAtom, SelectedMediaFolderSelector } from '../../state';
-import { FolderSelection } from './FolderSelection';
+import { FolderCreation } from './FolderCreation';
 import { LIMIT } from './Media';
 import { PaginationButton } from './PaginationButton';
 
@@ -82,7 +82,7 @@ export const Pagination: React.FunctionComponent<IPaginationProps> = ({}: React.
         </p>
       </div>
 
-      <FolderSelection />
+      <FolderCreation />
 
       <div className="flex justify-between sm:justify-end space-x-2 text-sm">
         <PaginationButton